Revert "Fixes max2code and mi2code assuming inputs are real scalars"
This reverts commit db3fec1d0e266cb79b8838bf49b4e8aae04a80d4.

+++ src/psyclone/tests/psyir/transformations/intrinsics/max2code_trans_test.py
@@ -69,8 +69,8 @@ def test_apply(fortran_reader, fortran_writer):
   integer :: j
   integer :: k
   integer :: l
-  integer :: res_max
-  integer :: tmp_max
+  real :: res_max
+  real :: tmp_max
 
   res_max = i
   tmp_max = j
